Various cleanups, oix now uses compiler warnings

+++ base/oix/rdebug.r
@@ -345,7 +345,7 @@ int get_name(dptr dp1, dptr dp0)
                             blkptr = blkptr->lelem.listprev;
                             i += blkptr->lelem.nused;
                         }
-                        sprintf(sbuf,"list#%d[%ld]",
+                        sprintf(sbuf,"list#%ld[%ld]",
                                 (long)blkptr->lelem.listprev->list.id, (long)i);
                         i = strlen(sbuf);
                         MemProtect(StrLoc(*dp0) = alcstr(sbuf,i));
@@ -354,8 +354,8 @@ int get_name(dptr dp1, dptr dp0)
                     case T_Record: { 		/* record */
                         struct b_constructor *c = blkptr->record.constructor;
                         i = varptr - blkptr->record.fields;
-                        sprintf(sbuf,"record %.*s#%d.%.*s", StrLen(c->name), StrLoc(c->name),
-                                blkptr->record.id,
+                        sprintf(sbuf,"record %.*s#%ld.%.*s", StrLen(c->name), StrLoc(c->name),
+                                (long)blkptr->record.id,
                                 StrLen(c->field_names[i]), StrLoc(c->field_names[i]));
                         i = strlen(sbuf);
                         MemProtect(StrLoc(*dp0) = alcstr(sbuf,i));
@@ -365,8 +365,8 @@ int get_name(dptr dp1, dptr dp0)
                     case T_Object: { 		/* object */
                         struct b_class *c = blkptr->object.class;
                         i = varptr - blkptr->object.fields;
-                        sprintf(sbuf,"object %.*s#%d.%.*s", StrLen(c->name), StrLoc(c->name),
-                                blkptr->object.id,
+                        sprintf(sbuf,"object %.*s#%ld.%.*s", StrLen(c->name), StrLoc(c->name),
+                                (long)blkptr->object.id,
                                 StrLen(c->fields[i]->name), StrLoc(c->fields[i]->name));
                         i = strlen(sbuf);
                         MemProtect(StrLoc(*dp0) = alcstr(sbuf,i));
@@ -413,7 +413,7 @@ static int keyref(bp, dp)
     else
         while(BlkType(bp) == T_Telem)
             bp = bp->telem.clink;
-        sprintf(sbuf, "table#%d[", bp->table.id);
+    sprintf(sbuf, "table#%ld[", (long)bp->table.id);
     { char * dest = sbuf + strlen(sbuf);
         strncpy(dest, s2, len);
         dest[len] = '\0';
